refactor: move TSqlLanguageService from dbflux_core to dbflux_driver_mssql
Dialect-specific language service implementations belong in their driver
crate, not in core. This mirrors the MySQL pattern established in PR #128.

- Add crates/dbflux_driver_mssql/src/language_service.rs with
  TSqlLanguageService + 3 unit tests (validate, editor_diagnostics,
  detect_dangerous).
- Wire the new module in dbflux_driver_mssql/src/lib.rs with pub mod +
  pub use re-export.
- Switch driver.rs:1394 from &dbflux_core::TSqlLanguageService to
  &crate::TSqlLanguageService.
- Remove TSqlLanguageService struct/impl from dbflux_core language_service.rs.
- Remove TSqlLanguageService from query/mod.rs and lib.rs re-export lists.

Closes #129
