+++ crates/dbflux_core/src/query/mod.rs
@@ -13,8 +13,8 @@ pub use generator::{
 };
 pub use language_service::{
     DangerousQueryKind, Diagnostic, DiagnosticSeverity, EditorDiagnostic, LanguageService,
-    SqlLanguageService, TSqlLanguageService, TextPosition, TextPositionRange, TextRange,
-    ValidationResult, classify_query_for_language, detect_dangerous_mongo, detect_dangerous_query,
+    SqlLanguageService, TextPosition, TextPositionRange, TextRange, ValidationResult,
+    classify_query_for_language, detect_dangerous_mongo, detect_dangerous_query,
     detect_dangerous_redis, detect_dangerous_sql, strip_leading_comments,
 };
 pub use safety::{classify_query_for_governance, classify_sql_execution, is_safe_read_query};
