+++ crates/dbflux_driver_mssql/src/lib.rs
@@ -10,5 +10,7 @@
 )]
 
 pub mod driver;
+pub mod language_service;
 
 pub use driver::{METADATA, MssqlDriver};
+pub use language_service::TSqlLanguageService;
